Improved filtering and gained some performance
It's easier and better to just split on any character not found in any
word in the vocabulary used. I've computed the characters used in
nltk.corpus.words.words() and created a regex which is the inverse of
the characters found, I this pattern use to split the message into terms.